+++ lua/codecompanion/strategies/chat/agents/tools/insert_edit_into_file.lua
@@ -1,11 +1,13 @@
-local Path = require("plenary.path")
 local buffers = require("codecompanion.utils.buffers")
+local codecompanion = require("codecompanion")
 local config = require("codecompanion.config")
 local diff = require("codecompanion.strategies.chat.agents.tools.helpers.diff")
 local log = require("codecompanion.utils.log")
 local patch = require("codecompanion.strategies.chat.agents.tools.helpers.patch")
 local ui = require("codecompanion.utils.ui")
 local wait = require("codecompanion.strategies.chat.agents.tools.helpers.wait")
+
+local Path = require("plenary.path")
 
 local api = vim.api
 local fmt = string.format
@@ -148,6 +150,15 @@ local function edit_buffer(bufnr, chat_bufnr, action, output_handler, opts)
     -- Wait for the user to accept or reject the edit
     return wait.for_decision(diff_id, { "CodeCompanionDiffAccepted", "CodeCompanionDiffRejected" }, function(result)
       if result.accepted then
+        -- Save the buffer
+        pcall(function()
+          api.nvim_buf_call(bufnr, function()
+            vim.cmd("silent! w")
+          end)
+        end)
+        -- NOTE: This is required to ensure folding works for chat buffers that aren't visible
+        codecompanion.restore(chat_bufnr)
+
         return output_handler(success)
       end
       return output_handler({